+++ src/Directory.build.props
@@ -27,17 +27,17 @@
 	</PropertyGroup>
 
 	<ItemGroup Condition="'$(IsTestProject)' != 'true'">
-		<PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.1.1" PrivateAssets="All" />
+		<PackageReference Include="Microsoft.SourceLink.GitHub" Version="8.0.0" PrivateAssets="All" />
 	</ItemGroup>
 
 	<ItemGroup Condition="'$(IsTestProject)' == 'true'">
 		<PackageReference Include="ReactiveUI.Testing" Version="19.*" />
-		<PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.7.2" />
-		<PackageReference Include="xunit" Version="2.6.1" />
-		<PackageReference Include="xunit.runner.console" Version="2.6.1" />
-		<PackageReference Include="xunit.runner.visualstudio" Version="2.5.3" />
+		<PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.8.0" />
+		<PackageReference Include="xunit" Version="2.6.2" />
+		<PackageReference Include="xunit.runner.console" Version="2.6.2" />
+		<PackageReference Include="xunit.runner.visualstudio" Version="2.5.4" />
 		<PackageReference Include="Xunit.StaFact" Version="1.1.11" />
-		<PackageReference Include="Verify.Xunit" Version="22.1.4" />
+		<PackageReference Include="Verify.Xunit" Version="22.5.0" />
 		<PackageReference Include="FluentAssertions" Version="6.12.0" />
 		<PackageReference Include="PublicApiGenerator" Version="11.0.0" />
 		<PackageReference Include="coverlet.msbuild" Version="6.0.0" PrivateAssets="All" />
@@ -52,7 +52,7 @@
 	<ItemGroup>
 		<PackageReference Include="Nerdbank.GitVersioning" Version="3.6.133" PrivateAssets="all" />
 		<PackageReference Include="stylecop.analyzers" Version="1.2.0-beta.507" PrivateAssets="all" />
-		<PackageReference Include="Roslynator.Analyzers" Version="4.6.1" PrivateAssets="All" />
+		<PackageReference Include="Roslynator.Analyzers" Version="4.6.2" PrivateAssets="All" />
 	</ItemGroup>
 
 	<ItemGroup>
