+++ src/Fusillade.Tests/API/ApiApprovalTests.FusilladeTests.DotNet7_0.verified.txt
@@ -1,4 +1,4 @@
-﻿[assembly: System.Runtime.Versioning.TargetFramework(".NETStandard,Version=v2.0", FrameworkDisplayName=".NET Standard 2.0")]
+﻿[assembly: System.Runtime.Versioning.TargetFramework(".NETCoreApp,Version=v7.0", FrameworkDisplayName=".NET 7.0")]
 namespace Fusillade
 {
     public interface IRequestCache
