+++ src/global.json
@@ -1,6 +1,6 @@
 {
     "sdk": {
-        "version": "7.0.400",
+        "version": "8.0.10",
         "rollForward": "latestMinor"
     },
     "msbuild-sdks": {
